requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
pendulum==3.0.0
requests==2.32.3
apache-airflow-providers-amazon==8.24.0
apache-airflow-providers-snowflake==5.5.2
astronomer-cosmos==1.5.1
protobuf